requirements.txt
crewai==0.141.0
python-dotenv==1.1.0
langchain-google-genai==2.0.10
pydantic==2.11.5
